moved creator pages to the CreatorHub dir . and updated the navbar and e2e to handle those changes
