+++ frontend/cypress/e2e/podcast/Episode.cy.tsx
@@ -35,7 +35,7 @@ describe("Episode_Create", () => {
       paths.never_gonna_give_you_up,
       "f2",
     );
-    cy.url().should("include", "/AddEpisode");
+    cy.url().should("include", "CreatorHub/AddEpisode");
     cy.contains(
       "An episode with the same name already exists for this podcast.",
     ).should("exist");
@@ -46,7 +46,7 @@ describe("Episode_Create", () => {
     cy.wait(500);
     cy.get('button[aria-label="loggedInMenu"]').click({ timeout: 5000 });
     cy.get("button").contains("My Podcasts").click({ timeout: 5000 });
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.wait(250);
     cy.get("[data-cy=podcast-image-aaaaaaaaaaaaaaaaaaaaaaaaa").click();
     cy.get("[data-cy=podcast-image-f2-legends").click();
@@ -58,7 +58,7 @@ describe("Episode_Create", () => {
       "Romain Grosjean",
     );
     cy.contains("Button", "Update").click();
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.wait(250);
     cy.get("[data-cy=podcast-image-aaaaaaaaaaaaaaaaaaaaaaaaa").click();
     cy.get("[data-cy=podcast-image-f2-legends").click();
@@ -74,7 +74,7 @@ describe("Episode_Create", () => {
       paths.never_gonna_give_you_up,
       "f2",
     );
-    cy.url().should("include", "/AddEpisode");
+    cy.url().should("include", "CreatorHub/AddEpisode");
     cy.contains("Required.").should("exist");
   });
 
@@ -87,20 +87,20 @@ describe("Episode_Create", () => {
       paths.never_gonna_give_you_up,
       null,
     );
-    cy.url().should("include", "/AddEpisode");
+    cy.url().should("include", "CreatorHub/AddEpisode");
     cy.contains("Please select the Podcast you wish to upload to");
   });
 
   // Users should be allowed to delete their own episodes
   it("Should successfully delete an episode", function () {
-    cy.visit("/MyPodcasts"); //Fix this when navbar gets more stable
-    cy.url().should("include", "/MyPodcasts");
+    cy.visit("CreatorHub/MyPodcasts"); //Fix this when navbar gets more stable
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.wait(250);
     cy.get("[data-cy=podcast-image-aaaaaaaaaaaaaaaaaaaaaaaaa").click();
     cy.get("[data-cy=podcast-image-f2-legends").click();
     cy.get("[data-cy=delete-button]").first().click();
     cy.contains("Button", "Delete").click({ timeout: 5000 });
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
   });
 
   // Episode names should be able to include special symbols not bound to ASCII characters
@@ -113,7 +113,7 @@ describe("Episode_Create", () => {
       "f2",
     );
     cy.get("button").contains("Finish").click({ timeout: 5000 });
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.wait(250);
     cy.get("[data-cy=podcast-image-aaaaaaaaaaaaaaaaaaaaaaaaa").click();
     cy.get("[data-cy=podcast-image-f2-legends").click();
@@ -123,7 +123,7 @@ describe("Episode_Create", () => {
 
   //Should add an episode from the podcast interface
   it("Should add an episode from the Podcast interface", () => {
-    cy.visit("/MyPodcasts"); //Fix this when navbar gets more stable
+    cy.visit("CreatorHub/MyPodcasts"); //Fix this when navbar gets more stable
     cy.get("button").contains("New Episode").click({ timeout: 5000 });
     cy.get('input[type="file"]').attachFile(paths.Episode_cover_science);
     cy.get('input[placeholder="Enter episode name..."]').type(
@@ -141,7 +141,7 @@ describe("Episode_Create", () => {
     cy.get("button[id=createBtn]").click();
     cy.get("button").contains("Finish").click({ timeout: 5000 });
     cy.wait(250);
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.wait(250);
     cy.get("[data-cy=podcast-image-aaaaaaaaaaaaaaaaaaaaaaaaa").click();
     cy.get("[data-cy=podcast-image-f2-legends").click();
@@ -159,7 +159,7 @@ describe("Episode_Create", () => {
       "f2",
     );
     cy.get("button").contains("Finish").click({ timeout: 5000 });
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.wait(250);
     cy.get("[data-cy=podcast-image-aaaaaaaaaaaaaaaaaaaaaaaaa").click();
     cy.get("[data-cy=podcast-image-f2-legends").click();
@@ -173,7 +173,7 @@ describe("Episode_Create", () => {
       "Cool pets",
       "A podcast about pets and their coolness.",
     );
-    cy.url().should("include", "/CreatePodcast");
+    cy.url().should("include", "CreatorHub/CreatePodcast");
     cy.contains("Cool pets");
     cy.episode_create(
       paths.shiba,
@@ -183,7 +183,7 @@ describe("Episode_Create", () => {
       "pets",
     );
     cy.get("button").contains("Finish").click({ timeout: 5000 });
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.episode_create(
       paths.cat,
       "Funny Cats",
@@ -192,14 +192,14 @@ describe("Episode_Create", () => {
       "pets",
     );
     cy.get("button").contains("Finish").click({ timeout: 5000 });
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.get("[data-cy=podcast-image-f2-legends").click();
     cy.get("[data-cy=podcast-image-cool-pets").click();
     cy.contains("Funny Cats");
     cy.contains("Silly cats");
     cy.get('[data-cy="podcast-delete"]').click();
     cy.contains("Button", "Delete").click();
-    cy.url().should("include", "/MyPodcasts");
+    cy.url().should("include", "CreatorHub/MyPodcasts");
     cy.wait(250);
     cy.get('[data-cy="podcast-image-cool-pets"]').should("not.exist");
     cy.get("Funny cats").should("not.exist");
